requirements.txt
torch>=1.8.1
torchvision>=0.10.0
numpy
opencv-python
pillow
tqdm
openai>=1.0.0
pyyaml
cython
cython-bbox
transformers
addict
yapf
timm
supervision>=0.22.0
pycocotools
loguru
scikit-image
Pillow
thop
ninja
tabulate
tensorboard
lap
motmetrics
filterpy
h5py